requirements.txt
Django==5.2.6
djangorestframework==3.16.1
django-cors-headers==4.9.0
psycopg2-binary==2.9.10
python-decouple==3.8
whitenoise==6.11.0
dj-database-url==3.0.1
Pillow==11.3.0
qrcode[pil]==8.2
openpyxl==3.1.5
pandas==2.3.2
matplotlib==3.10.6
seaborn==0.13.2
plotly==6.3.0
scipy==1.16.2
numpy==2.3.3
gunicorn==23.0.0

reportlab==4.2.5
